refactor: useCouponDiscount에서  buyXgetY 할인 로직을 위한 계산 함수 분리
