+++ src/hooks/useCouponDiscount.ts
@@ -8,7 +8,7 @@ import {
   SHIPPING_FEE,
   REMOTE_AREA_SHIPPING_FEE,
 } from "../constants";
-import { CartItem } from "../types/type";
+import { calculateBuyXGetYDiscount } from "../utils/discounts";
 
 export const useCouponDiscount = () => {
   const { appliedCoupons } = useCouponContext();
@@ -67,32 +67,4 @@ export const useCouponDiscount = () => {
     couponDiscount,
     orderPrice,
   };
-};
-
-const calculateBuyXGetYDiscount = (
-  selectedItems: CartItem[],
-  coupon: Coupon
-): number => {
-  if (!coupon.buyQuantity || !coupon.getQuantity) {
-    return 0;
-  }
-
-  const requiredQuantity = coupon.buyQuantity + coupon.getQuantity;
-
-  const eligibleItems = selectedItems.filter(
-    (item) => item.quantity >= requiredQuantity
-  );
-
-  if (eligibleItems.length === 0) {
-    return 0;
-  }
-
-  let maxPrice = 0;
-  eligibleItems.forEach((item) => {
-    if (item.product.price > maxPrice) {
-      maxPrice = item.product.price;
-    }
-  });
-
-  return maxPrice;
 };
